Address feedback
Use shared files; Use all crashes
