--- SasquatchMac/SasquatchMacSwift-Preview Today Extension/Base.lproj/TodayViewController.xib
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
-    <dependencies>
-        <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="14460.31"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="TodayViewController" customModule="SasquatchMacSwift_Preview_Today_Extension" customModuleProvider="target">
-            <connections>
-                <outlet property="extensionLabel" destination="1sU-Wy-afS" id="9hU-GX-oLr"/>
-                <outlet property="view" destination="Hz6-mo-xeY" id="0bl-1N-x8E"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customView simulatedAppContext="notificationCenter" translatesAutoresizingMaskIntoConstraints="NO" id="Hz6-mo-xeY">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="64"/>
-            <subviews>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="1sU-Wy-afS">
-                    <rect key="frame" x="6" y="40" width="363" height="16"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" alignment="center" title="Run #000000" id="bTn-1C-B7R">
-                        <font key="font" metaFont="cellTitle"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <button verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="Hzq-IV-DAH" userLabel="CrashMe">
-                    <rect key="frame" x="2" y="4" width="371" height="32"/>
-                    <buttonCell key="cell" type="push" title="Crash Me!" bezelStyle="rounded" alignment="center" borderStyle="border" imageScaling="proportionallyDown" inset="2" id="3hJ-SQ-2CW">
-                        <behavior key="behavior" pushIn="YES" lightByBackground="YES" lightByGray="YES"/>
-                        <font key="font" metaFont="system"/>
-                    </buttonCell>
-                    <connections>
-                        <action selector="crashMe:" target="-2" id="ypo-ef-jiW"/>
-                    </connections>
-                </button>
-            </subviews>
-            <constraints>
-                <constraint firstItem="1sU-Wy-afS" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="8" id="0za-kV-ANi"/>
-                <constraint firstItem="Hzq-IV-DAH" firstAttribute="leading" secondItem="1sU-Wy-afS" secondAttribute="leading" id="AvQ-s8-PYf"/>
-                <constraint firstItem="1sU-Wy-afS" firstAttribute="top" secondItem="Hz6-mo-xeY" secondAttribute="top" constant="8" id="FhG-Ze-3qQ"/>
-                <constraint firstAttribute="trailing" secondItem="1sU-Wy-afS" secondAttribute="trailing" constant="8" id="Gee-EA-ysn"/>
-                <constraint firstItem="Hzq-IV-DAH" firstAttribute="trailing" secondItem="1sU-Wy-afS" secondAttribute="trailing" id="oHE-TF-z9T"/>
-                <constraint firstItem="Hzq-IV-DAH" firstAttribute="top" secondItem="1sU-Wy-afS" secondAttribute="bottom" constant="8" id="sKq-Eb-RdL"/>
-                <constraint firstAttribute="bottom" secondItem="Hzq-IV-DAH" secondAttribute="bottom" constant="11" id="u78-xO-zFN"/>
-            </constraints>
-            <point key="canvasLocation" x="138.5" y="154"/>
-        </customView>
-    </objects>
-</document>
